Broke out handleTracklistPostResponse from within tracklistForm listener.
