+++ client/listeners/tracklistForm.js
@@ -1,15 +1,13 @@
-const { addLoadRef, setOverwriteRef } = require('./../utils/elementRefs')
+const { addLoadRef } = require('./../utils/elementRefs')
 const state = require('./../state/state.js')
 
-const submitTracklist = require('./../utils/submitTracklist.js')
-const sendTracklistPostRequest = require('./../utils/sendTracklistPostRequest.js')
-
-const { attachOnZipListener } = require('./socketListeners')
+const submitTracklist = require('./../utils/submitTracklist')
+const sendTracklistPostRequest = require('./../utils/sendTracklistPostRequest')
+const handleTracklistPostResponse = require('./../utils/handleTracklistPostResponse')
 
 function attachTracklistFormListener() {
   const $tracklistForm = addLoadRef('tracklist-form')
   const $tracklistError = addLoadRef('tracklist-error-message-container')
-  const $trackFinalContainer = addLoadRef('track-final-container')
   const $downloadProgress = addLoadRef('album-download-progress')
   const $sliceProgress = addLoadRef('track-slice-progress')
 
@@ -39,33 +37,7 @@ function attachTracklistFormListener() {
       }, 2000)
     }
 
-    sendTracklistPostRequest(tracklistPost).then(response => {
-      const $audioPlayer = addLoadRef('audio-player')
-      const $nowPlaying = addLoadRef('now-playing')
-      if (response.status === 202) {
-        $trackFinalContainer.addEventListener('click', e => {
-          state.selectedTrack = state.tracklist[(parseInt(e.target.dataset.tracknum, 10) - 1)]
-          if (!state.selectedTrack) return
-          state.audio.selectTrack(state.selectedTrack)
-          const $selected = addLoadRef(`track-final-${e.target.dataset.tracknum}`)
-          for (let i = 1; i <= state.tracklist.length; i++) {
-            const $track = setOverwriteRef(`track-final-${i}`)
-            if ($track.classList.value.includes('selected')) $track.classList.remove('selected')
-          }
-          $selected.setAttribute('class', 'track-final selected')
-
-          $nowPlaying.textContent = state.selectedTrack.trackName
-          const trackFileName = state.selectedTrack.trackName.split(' ').join('-')
-
-          const trackPath = '/download/' + state.albumMetadata.videoId + '/tracks/' + state.socketId + '/' + trackFileName + '.mp3'
-          $audioPlayer.pause()
-          $audioPlayer.src = trackPath
-          $audioPlayer.play()
-        })
-        attachOnZipListener()
-      }
-      else (console.log('Tracklist request failed.'))
-    })
+    sendTracklistPostRequest(tracklistPost).then(handleTracklistPostResponse)
   })
 }
